Move accessible text to a hidden element
Using aria labels was fine for screen readers, but doesn't cater for a
core experience where the state icons may not be visible. The state text
has now moved to a hidden element so that the core experience (and
failed CSS experience) are more usable.

Before
------

  - **No CSS:** Displays as an ordered list without state icons and text
  - **IE < 9:** Displays as a stacked list without state icons and text
  - **IE 9:** Displays as a stacked list with icons but no text
  - **IE 10+:** Displays as a horizontal list with icons and no text
  - **Others:** Displays as a horizontal list with icons and no text

After
-----

  - **No CSS:** Displays as an ordered list with text but no icons
  - **IE < 9:** Displays as a stacked list with text but no icons
  - **IE 9:** Displays as a stacked list with text and icons
  - **IE 10+:** Displays as a horizontal list with icons and no text
  - **Others:** Displays as a horizontal list with icons and no text

The only compromise here is that IE 9 gets both text and icons, but the
code hoops I'd have to jump through to hide the text would be even more
gross than what I've had to do already.